requirements/requirements.txt
# Storage + serialization
msgpack
zstandard
pyarrow
polars>=0.20.6
numpy
xmltodict

#LLM + vector
google-generativeai
hnswlib
torch
transformers
sentence-transformers[torch]>=2.5.0 


# visualization
plotly

# Async operations
aiofiles
aiohttp

# VCS
diff-match-patch 
patience

# Monitering
langfuse

# Files
pyyaml>=6.0.1
xmltodict>=0.13.0
